Conflicts:
	Turbulence/WebSite/App_Code/WebService.cs

+++ Turbulence/Turbulence/SQLInterface/workers/GetChannelPressure.cs
@@ -54,12 +54,12 @@ namespace Turbulence.SQLInterface.workers
             };
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
         {
             throw new NotImplementedException();
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
         {
             return CalcLagInterpolation(blob, input);
         }
@@ -76,9 +76,9 @@ namespace Turbulence.SQLInterface.workers
         /// The Lagrangian evaluation function [LagInterpolation.EvaluateOpt] was moved
         /// into the function and some loop unrolling was performed.
         /// </remarks>
-        unsafe public float[] CalcLagInterpolation(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
+        unsafe public double[] CalcLagInterpolation(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
         {
-            float[] up = new float[1]; // Result value for the user
+            double[] up = new double[1]; // Result value for the user
 
             if (spatialInterp == TurbulenceOptions.SpatialInterpolation.None)
             {
@@ -177,7 +177,7 @@ namespace Turbulence.SQLInterface.workers
                             double a = lagint[2 * kernelSize + iLagIntz + iz - startz];
                             a1 += b1 * a;
                         }
-                        up[0] = (float)a1;
+                        up[0] = a1;
                     }
                 }
             }
