+++ Turbulence/Turbulence/SQLInterface/workers/GetPressureGradient.cs
@@ -48,15 +48,15 @@ namespace Turbulence.SQLInterface.workers
                 new SqlMetaData("dpdz", SqlDbType.Real) };
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
         {
             //return new float[] { 5.0f, 5.0f, 5.0f, 5.0f, 5.0f };
-            
-            return (float[]) compute.CalcPressureGradient(blob,
+
+            return (double[])compute.CalcPressureGradient(blob,
                 new float[] { (float)input.x, (float)input.y, (float)input.z }, spatialInterp).Clone();
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
         {
             throw new NotImplementedException();
         }
